% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcanvas-package.R
\docType{package}
\name{rcanvas-package}
\alias{rcanvas}
\alias{rcanvas-package}
\title{rcanvas: Work with Canvas Courses using the Canvas API}
\description{
Retrieve and modify Canvas data via the 'Canvas' \url{https://canvas.instructure.com/doc/api/index.html} API.
}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
\keyword{internal}
