% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas-api-variables.R
\name{set_canvas_domain}
\alias{set_canvas_domain}
\title{Set Canvas Domain}
\usage{
set_canvas_domain(domain = "https://canvas.instructure.com")
}
\arguments{
\item{domain}{Canvas Domain. Defaults to \url{https://canvas.instructure.com}.}
}
\description{
Set the Canvas domain to be used in the current session.
}
\examples{
\dontrun{
set_canvas_domain(domain = "https://canvas.instructure.com")
}
}
