% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas-api-variables.R
\name{set_canvas_token}
\alias{set_canvas_token}
\title{Set Canvas API Key}
\usage{
set_canvas_token(api_key = NULL, ask = is_interactive())
}
\arguments{
\item{api_key}{Canvas Web API Key}

\item{ask}{Try to set the API key if not found. Requires session to be
interactive.}
}
\description{
Set the Canvas API Key to be used in the current session.
}
\examples{
\dontrun{
set_canvas_token(api_key = "my-secret-1823510823uadf-api-key")
}
}
